refactor: isolate archive verification
